script.js
you have a very cool project, I hope with these examples I have expanded your horizons a little on the options for an approach to work, with the help of such methods, you can add any number of questions! Good luck with your studies, you'll be able to take on a bigger task and you'll be strong!